+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/SentryPlugin.kt
@@ -23,6 +23,9 @@ import org.gradle.api.Task
 import org.gradle.api.plugins.ExtraPropertiesExtension
 
 class SentryPlugin : Plugin<Project> {
+
+    private val sep = File.separator
+
     override fun apply(project: Project) {
         val extension = project.extensions.create(
             "sentry",
@@ -52,12 +55,11 @@ class SentryPlugin : Plugin<Project> {
                 val sentryProperties = getPropertiesFilePath(project, variant)
 
                 val isMinifyEnabled = variant.buildType.isMinifyEnabled
+                val isDebuggable = variant.buildType.isDebuggable
 
                 var preBundleTask: Task? = null
                 var transformerTask: Task? = null
                 var packageBundleTask: Task? = null
-
-                val sep = File.separator
 
                 if (isMinifyEnabled) {
                     preBundleTask = withLogging(project.logger, "preBundleTask") {
@@ -127,25 +129,26 @@ class SentryPlugin : Plugin<Project> {
                     packageBundleTask?.dependsOn(uploadSentryProguardMappingsTask)
                 }
 
-                // Setup the task to upload native symbols task after the assembling task
-                val uploadNativeSymbolsTask = project.tasks.register(
-                    "uploadNativeSymbolsFor$taskSuffix",
-                    SentryUploadNativeSymbolsTask::class.java
-                ) {
-                    it.workingDir(project.rootDir)
-                    it.cliExecutable.set(cliExecutable)
-                    it.sentryProperties.set(
-                        sentryProperties?.let { file -> project.file(file) }
-                    )
-                    it.includeNativeSources.set(extension.includeNativeSources.get())
-                    it.variantName.set(variant.name)
-                    it.sentryOrganization.set(sentryOrgParameter)
-                    it.sentryProject.set(sentryProjectParameter)
-                }
-
+                // only debug symbols of non debuggable code should be uploaded (aka release builds).
                 // uploadNativeSymbolsTask will only be executed after the assemble task
                 // and also only if `uploadNativeSymbols` is enabled, as this is an opt-in feature.
-                if (extension.uploadNativeSymbols.get()) {
+                if (!isDebuggable && extension.uploadNativeSymbols.get()) {
+                    // Setup the task to upload native symbols task after the assembling task
+                    val uploadNativeSymbolsTask = project.tasks.register(
+                        "uploadNativeSymbolsFor$taskSuffix",
+                        SentryUploadNativeSymbolsTask::class.java
+                    ) {
+                        it.workingDir(project.rootDir)
+                        it.cliExecutable.set(cliExecutable)
+                        it.sentryProperties.set(
+                            sentryProperties?.let { file -> project.file(file) }
+                        )
+                        it.includeNativeSources.set(extension.includeNativeSources.get())
+                        it.variantName.set(variant.name)
+                        it.sentryOrganization.set(sentryOrgParameter)
+                        it.sentryProject.set(sentryProjectParameter)
+                    }
+
                     getAssembleTaskProvider(variant)?.configure {
                         it.finalizedBy(
                             uploadNativeSymbolsTask